requirements.txt
streamlit
openai-whisper
torch
sounddevice
numpy
python-docx